break SusyNt down into one file per class
For now we still allow to include SusyNt, but at some point
we will want to force the Selector to include only what they need.
